change catalog's Blocks and Limit data type from Double to BigDecimal
affected classes and attributes described in this comment: https://github.com/killbill/killbill/issues/1469#issuecomment-1028580177

+++ src/main/java/org/killbill/billing/catalog/api/Limit.java
@@ -16,13 +16,15 @@
 
 package org.killbill.billing.catalog.api;
 
+import java.math.BigDecimal;
+
 public interface Limit {
 
     public Unit getUnit();
 
-    public Double getMax();
+    public BigDecimal getMax();
 
-    public Double getMin();
+    public BigDecimal getMin();
 
     public boolean compliesWith(double value);
